# docs/swagger/insurance.yaml
openapi: 3.0.3
info:
  title: Insurance Asset API
  version: 1.0.0
paths:
  /insurance/customers/{customerId}/assets:
    get:
      summary: Get insurance assets for a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsuranceAssetResponse'
              examples:
                sample:
                  value:
                    customerId: "123"
                    insuranceAssets:
                      - policyNumber: LIF-1000
                        policyType: life
                        coverage: 100000
                        premiumStatus: paid
                        currency: TWD
                        lastUpdated: "2025-10-20T02:34:56Z"
                      - policyNumber: HLT-2000
                        policyType: health
                        coverage: 50000
                        premiumStatus: due
                        currency: TWD
                        lastUpdated: "2025-10-20T02:35:10Z"
                    totalCoverage: 150000
                    currency: TWD
                    traceId: "c9e8f7g6-..."
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Downstream unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    InsuranceAssetResponse:
      type: object
      required: [customerId, insuranceAssets, totalCoverage, currency, traceId]
      properties:
        customerId:
          type: string
        insuranceAssets:
          type: array
          items:
            $ref: '#/components/schemas/InsuranceAssetItem'
        totalCoverage:
          type: number
          format: double
        currency:
          type: string
        traceId:
          type: string
    InsuranceAssetItem:
      type: object
      required: [policyNumber, policyType, coverage, premiumStatus, currency, lastUpdated]
      properties:
        policyNumber:
          type: string
          description: Unique policy identifier generated by the insurance module
        policyType:
          type: string
        coverage:
          type: number
          format: double
        premiumStatus:
          type: string
        currency:
          type: string
        lastUpdated:
          type: string
          format: date-time
    ErrorResponse:
      $ref: './bank.yaml#/components/schemas/ErrorResponse'